requirements.txt
flask==2.2.5
requests==2.31.0
beautifulsoup4==4.12.3
pymysql==1.1.0
pandas==2.2.2
apache-airflow==2.8.1